Refactored candidates
