+++ src/ipc/CMakeLists.txt
@@ -17,6 +17,7 @@ target_sources(ipc_toolkit PRIVATE ${SOURCES})
 add_subdirectory(barrier)
 add_subdirectory(broad_phase)
 add_subdirectory(ccd)
+add_subdirectory(candidates)
 add_subdirectory(collisions)
 add_subdirectory(distance)
 add_subdirectory(friction)
