--- src/ipc/broad_phase/collision_candidate.hpp
@@ -1,228 +1,0 @@
-#pragma once
-
-#include <ipc/ccd/ccd.hpp>
-
-#include <Eigen/Core>
-
-#include <vector>
-
-namespace ipc {
-
-/// Virtual class for candidates that support CCD.
-struct ContinuousCollisionCandidate {
-    virtual ~ContinuousCollisionCandidate() { }
-
-    /// Perform narrow-phase CCD on the candidate.
-    /// @param[in] V0 Mesh vertex positions at the start of the time step.
-    /// @param[in] V1 Mesh vertex positions at the end of the time step.
-    /// @param[in] E Mesh edges as rows of indicies into V.
-    /// @param[in] F Mesh triangular faces as rows of indicies into V.
-    /// @param[out] toi Computed time of impact (normalized).
-    /// @param[in] tmax Maximum time (normalized) to look for collisions. Should be in [0, 1].
-    /// @param[in] tolerance CCD tolerance used by Tight-Inclusion CCD.
-    /// @param[in] max_iterations Maximum iterations used by Tight-Inclusion CCD.
-    /// @param[in] conservative_rescaling Conservative rescaling value used to avoid taking steps exactly to impact.
-    /// @return If the candidate had a collision over the time interval.
-    virtual bool
-    ccd(const Eigen::MatrixXd& V0,
-        const Eigen::MatrixXd& V1,
-        const Eigen::MatrixXi& E,
-        const Eigen::MatrixXi& F,
-        double& toi,
-        const double tmax = 1.0,
-        const double tolerance = DEFAULT_CCD_TOLERANCE,
-        const long max_iterations = DEFAULT_CCD_MAX_ITERATIONS,
-        const double conservative_rescaling =
-            DEFAULT_CCD_CONSERVATIVE_RESCALING) const = 0;
-
-    // Print the vertices of the CCD query for debugging.
-    virtual void print_ccd_query(
-        const Eigen::MatrixXd& V0,
-        const Eigen::MatrixXd& V1,
-        const Eigen::MatrixXi& E,
-        const Eigen::MatrixXi& F) const = 0;
-};
-
-struct VertexVertexCandidate {
-    VertexVertexCandidate(long vertex0_index, long vertex1_index);
-
-    bool operator==(const VertexVertexCandidate& other) const;
-    bool operator!=(const VertexVertexCandidate& other) const
-    {
-        return !(*this == other);
-    }
-
-    /// @brief Compare EdgeVertexCandidates for sorting.
-    bool operator<(const VertexVertexCandidate& other) const;
-
-    long vertex0_index;
-    long vertex1_index;
-};
-
-struct EdgeVertexCandidate : ContinuousCollisionCandidate {
-    EdgeVertexCandidate(long edge_index, long vertex_index);
-
-    bool operator==(const EdgeVertexCandidate& other) const;
-    bool operator!=(const EdgeVertexCandidate& other) const
-    {
-        return !(*this == other);
-    }
-
-    /// @brief Compare EdgeVertexCandidates for sorting.
-    bool operator<(const EdgeVertexCandidate& other) const;
-
-    /// Perform narrow-phase CCD on the candidate.
-    /// @param[in] V0 Mesh vertex positions at the start of the time step.
-    /// @param[in] V1 Mesh vertex positions at the end of the time step.
-    /// @param[in] E Mesh edges as rows of indicies into V.
-    /// @param[in] F Mesh triangular faces as rows of indicies into V.
-    /// @param[out] toi Computed time of impact (normalized).
-    /// @param[in] tmax Maximum time (normalized) to look for collisions. Should be in [0, 1].
-    /// @param[in] tolerance CCD tolerance used by Tight-Inclusion CCD.
-    /// @param[in] max_iterations Maximum iterations used by Tight-Inclusion CCD.
-    /// @param[in] conservative_rescaling Conservative rescaling value used to avoid taking steps exactly to impact.
-    /// @return If the candidate had a collision over the time interval.
-    bool
-    ccd(const Eigen::MatrixXd& V0,
-        const Eigen::MatrixXd& V1,
-        const Eigen::MatrixXi& E,
-        const Eigen::MatrixXi& F,
-        double& toi,
-        const double tmax = 1.0,
-        const double tolerance = DEFAULT_CCD_TOLERANCE,
-        const long max_iterations = DEFAULT_CCD_MAX_ITERATIONS,
-        const double conservative_rescaling =
-            DEFAULT_CCD_CONSERVATIVE_RESCALING) const override;
-
-    void print_ccd_query(
-        const Eigen::MatrixXd& V0,
-        const Eigen::MatrixXd& V1,
-        const Eigen::MatrixXi& E,
-        const Eigen::MatrixXi& F) const override;
-
-    long edge_index;
-    long vertex_index;
-};
-
-struct EdgeEdgeCandidate : ContinuousCollisionCandidate {
-    EdgeEdgeCandidate(long edge0_index, long edge1_index);
-
-    bool operator==(const EdgeEdgeCandidate& other) const;
-    bool operator!=(const EdgeEdgeCandidate& other) const
-    {
-        return !(*this == other);
-    }
-
-    /// @brief Compare EdgeEdgeCandidates for sorting.
-    bool operator<(const EdgeEdgeCandidate& other) const;
-
-    /// Perform narrow-phase CCD on the candidate.
-    /// @param[in] V0 Mesh vertex positions at the start of the time step.
-    /// @param[in] V1 Mesh vertex positions at the end of the time step.
-    /// @param[in] E Mesh edges as rows of indicies into V.
-    /// @param[in] F Mesh triangular faces as rows of indicies into V.
-    /// @param[out] toi Computed time of impact (normalized).
-    /// @param[in] tmax Maximum time (normalized) to look for collisions. Should be in [0, 1].
-    /// @param[in] tolerance CCD tolerance used by Tight-Inclusion CCD.
-    /// @param[in] max_iterations Maximum iterations used by Tight-Inclusion CCD.
-    /// @param[in] conservative_rescaling Conservative rescaling value used to avoid taking steps exactly to impact.
-    /// @return If the candidate had a collision over the time interval.
-    bool
-    ccd(const Eigen::MatrixXd& V0,
-        const Eigen::MatrixXd& V1,
-        const Eigen::MatrixXi& E,
-        const Eigen::MatrixXi& F,
-        double& toi,
-        const double tmax = 1.0,
-        const double tolerance = DEFAULT_CCD_TOLERANCE,
-        const long max_iterations = DEFAULT_CCD_MAX_ITERATIONS,
-        const double conservative_rescaling =
-            DEFAULT_CCD_CONSERVATIVE_RESCALING) const override;
-
-    void print_ccd_query(
-        const Eigen::MatrixXd& V0,
-        const Eigen::MatrixXd& V1,
-        const Eigen::MatrixXi& E,
-        const Eigen::MatrixXi& F) const override;
-
-    long edge0_index;
-    long edge1_index;
-};
-
-/// @brief Candidate for <b>intersection</b> between edge and face.
-///
-/// Not included in Candidates because it is not a collision candidate.
-struct EdgeFaceCandidate {
-    EdgeFaceCandidate(long edge_index, long face_index);
-
-    bool operator==(const EdgeFaceCandidate& other) const;
-    bool operator!=(const EdgeFaceCandidate& other) const
-    {
-        return !(*this == other);
-    }
-
-    /// @brief Compare EdgeFaceCandidate for sorting.
-    bool operator<(const EdgeFaceCandidate& other) const;
-
-    long edge_index;
-    long face_index;
-};
-
-struct FaceVertexCandidate : ContinuousCollisionCandidate {
-    FaceVertexCandidate(long face_index, long vertex_index);
-
-    bool operator==(const FaceVertexCandidate& other) const;
-    bool operator!=(const FaceVertexCandidate& other) const
-    {
-        return !(*this == other);
-    }
-
-    /// @brief Compare FaceVertexCandidate for sorting.
-    bool operator<(const FaceVertexCandidate& other) const;
-
-    bool
-    ccd(const Eigen::MatrixXd& V0,
-        const Eigen::MatrixXd& V1,
-        const Eigen::MatrixXi& E,
-        const Eigen::MatrixXi& F,
-        double& toi,
-        const double tmax = 1.0,
-        const double tolerance = DEFAULT_CCD_TOLERANCE,
-        const long max_iterations = DEFAULT_CCD_MAX_ITERATIONS,
-        const double conservative_rescaling =
-            DEFAULT_CCD_CONSERVATIVE_RESCALING) const override;
-
-    void print_ccd_query(
-        const Eigen::MatrixXd& V0,
-        const Eigen::MatrixXd& V1,
-        const Eigen::MatrixXi& E,
-        const Eigen::MatrixXi& F) const override;
-
-    long face_index;
-    long vertex_index;
-};
-
-struct Candidates {
-    std::vector<EdgeVertexCandidate> ev_candidates;
-    std::vector<EdgeEdgeCandidate> ee_candidates;
-    std::vector<FaceVertexCandidate> fv_candidates;
-
-    Candidates() { }
-
-    size_t size() const;
-
-    bool empty() const;
-
-    void clear();
-
-    ContinuousCollisionCandidate& operator[](size_t idx);
-    const ContinuousCollisionCandidate& operator[](size_t idx) const;
-
-    bool save_obj(
-        const std::string& filename,
-        const Eigen::MatrixXd& V,
-        const Eigen::MatrixXi& E,
-        const Eigen::MatrixXi& F) const;
-};
-
-} // namespace ipc
