+++ src/ipc/candidates/edge_face.cpp
@@ -1,0 +1,29 @@
+#include "edge_face.hpp"
+
+namespace ipc {
+
+EdgeFaceCandidate::EdgeFaceCandidate(long edge_index, long face_index)
+    : edge_index(edge_index)
+    , face_index(face_index)
+{
+}
+
+bool EdgeFaceCandidate::operator==(const EdgeFaceCandidate& other) const
+{
+    return edge_index == other.edge_index && face_index == other.face_index;
+}
+
+bool EdgeFaceCandidate::operator!=(const EdgeFaceCandidate& other) const
+{
+    return !(*this == other);
+}
+
+bool EdgeFaceCandidate::operator<(const EdgeFaceCandidate& other) const
+{
+    if (edge_index == other.edge_index) {
+        return face_index < other.face_index;
+    }
+    return edge_index < other.edge_index;
+}
+
+} // namespace ipc
