+++ src/ipc/collisions/face_vertex.hpp
@@ -1,42 +1,57 @@
 #pragma once
 
-#include <ipc/broad_phase/collision_candidate.hpp>
+#include <ipc/candidates/face_vertex.hpp>
 #include <ipc/collisions/collision_constraint.hpp>
 #include <ipc/utils/eigen_ext.hpp>
 
 namespace ipc {
 
 struct FaceVertexConstraint : FaceVertexCandidate, CollisionConstraint {
-    FaceVertexConstraint(long face_index, long vertex_index);
-    FaceVertexConstraint(const FaceVertexCandidate& candidate);
+    using FaceVertexCandidate::FaceVertexCandidate;
 
-    int num_vertices() const override { return 4; };
+    int num_vertices() const override
+    {
+        return FaceVertexCandidate::num_vertices();
+    };
+
     std::array<long, 4> vertex_indices(
         const Eigen::MatrixXi& E, const Eigen::MatrixXi& F) const override
     {
-        return { { vertex_index, //
-                   F(face_index, 0), F(face_index, 1), F(face_index, 2) } };
+        return FaceVertexCandidate::vertex_indices(E, F);
     }
 
     double compute_distance(
         const Eigen::MatrixXd& V,
         const Eigen::MatrixXi& E,
-        const Eigen::MatrixXi& F) const override;
+        const Eigen::MatrixXi& F) const override
+    {
+        return FaceVertexCandidate::compute_distance(
+            V, E, F, PointTriangleDistanceType::P_T);
+    }
 
     VectorMax12d compute_distance_gradient(
         const Eigen::MatrixXd& V,
         const Eigen::MatrixXi& E,
-        const Eigen::MatrixXi& F) const override;
+        const Eigen::MatrixXi& F) const override
+    {
+        return FaceVertexCandidate::compute_distance_gradient(
+            V, E, F, PointTriangleDistanceType::P_T);
+    }
 
     MatrixMax12d compute_distance_hessian(
         const Eigen::MatrixXd& V,
         const Eigen::MatrixXi& E,
-        const Eigen::MatrixXi& F) const override;
+        const Eigen::MatrixXi& F) const override
+    {
+        return FaceVertexCandidate::compute_distance_hessian(
+            V, E, F, PointTriangleDistanceType::P_T);
+    }
 
     template <typename H>
     friend H AbslHashValue(H h, const FaceVertexConstraint& fv)
     {
-        return H::combine(std::move(h), fv.face_index, fv.vertex_index);
+        return AbslHashValue(
+            std::move(h), static_cast<const FaceVertexCandidate&>(fv));
     }
 };
