+++ src/ipc/distance/point_edge.hpp
@@ -13,24 +13,6 @@ namespace ipc {
 /// @param[in] p The point.
 /// @param[in] e0 The first vertex of the edge.
 /// @param[in] e1 The second vertex of the edge.
-/// @return The distance between the point and edge.
-template <typename DerivedP, typename DerivedE0, typename DerivedE1>
-auto point_edge_distance(
-    const Eigen::MatrixBase<DerivedP>& p,
-    const Eigen::MatrixBase<DerivedE0>& e0,
-    const Eigen::MatrixBase<DerivedE1>& e1)
-{
-    assert(p.size() == 2 || p.size() == 3);
-    assert(e0.size() == 2 || e0.size() == 3);
-    assert(e1.size() == 2 || e1.size() == 3);
-    return point_edge_distance(p, e0, e1, point_edge_distance_type(p, e0, e1));
-}
-
-/// @brief Compute the distance between a point and edge in 2D or 3D.
-/// @note The distance is actually squared distance.
-/// @param[in] p The point.
-/// @param[in] e0 The first vertex of the edge.
-/// @param[in] e1 The second vertex of the edge.
 /// @param[in] dtype The point edge distance type to compute.
 /// @return The distance between the point and edge.
 template <typename DerivedP, typename DerivedE0, typename DerivedE1>
@@ -38,7 +20,7 @@ auto point_edge_distance(
     const Eigen::MatrixBase<DerivedP>& p,
     const Eigen::MatrixBase<DerivedE0>& e0,
     const Eigen::MatrixBase<DerivedE1>& e1,
-    const PointEdgeDistanceType dtype)
+    const PointEdgeDistanceType dtype = PointEdgeDistanceType::AUTO)
 {
     assert(p.size() == 2 || p.size() == 3);
     assert(e0.size() == 2 || e0.size() == 3);
@@ -47,39 +29,21 @@ auto point_edge_distance(
     switch (dtype) {
     case PointEdgeDistanceType::P_E0:
         return point_point_distance(p, e0);
+
     case PointEdgeDistanceType::P_E1:
         return point_point_distance(p, e1);
+
     case PointEdgeDistanceType::P_E:
         return point_line_distance(p, e0, e1);
+
+    case PointEdgeDistanceType::AUTO:
+        return point_edge_distance(
+            p, e0, e1, point_edge_distance_type(p, e0, e1));
+
     default:
         throw std::invalid_argument(
             "Invalid distance type for point-edge distance!");
     }
-}
-
-/// @brief Compute the gradient of the distance between a point and edge.
-/// @note The distance is actually squared distance.
-/// @param[in] p The point.
-/// @param[in] e0 The first vertex of the edge.
-/// @param[in] e1 The second vertex of the edge.
-/// @param[out] grad Gradient of the distance wrt p, e0, and e1.
-template <
-    typename DerivedP,
-    typename DerivedE0,
-    typename DerivedE1,
-    typename DerivedGrad>
-void point_edge_distance_gradient(
-    const Eigen::MatrixBase<DerivedP>& p,
-    const Eigen::MatrixBase<DerivedE0>& e0,
-    const Eigen::MatrixBase<DerivedE1>& e1,
-    Eigen::PlainObjectBase<DerivedGrad>& grad)
-{
-    assert(p.size() == 2 || p.size() == 3);
-    assert(e0.size() == 2 || e0.size() == 3);
-    assert(e1.size() == 2 || e1.size() == 3);
-
-    return point_edge_distance_gradient(
-        p, e0, e1, point_edge_distance_type(p, e0, e1), grad);
 }
 
 /// @brief Compute the gradient of the distance between a point and edge.
@@ -98,8 +62,8 @@ void point_edge_distance_gradient(
     const Eigen::MatrixBase<DerivedP>& p,
     const Eigen::MatrixBase<DerivedE0>& e0,
     const Eigen::MatrixBase<DerivedE1>& e1,
-    const PointEdgeDistanceType dtype,
-    Eigen::PlainObjectBase<DerivedGrad>& grad)
+    Eigen::PlainObjectBase<DerivedGrad>& grad,
+    const PointEdgeDistanceType dtype = PointEdgeDistanceType::AUTO)
 {
     int dim = p.size();
     assert(e0.size() == dim);
@@ -125,35 +89,15 @@ void point_edge_distance_gradient(
         point_line_distance_gradient(p, e0, e1, grad);
         break;
 
+    case PointEdgeDistanceType::AUTO:
+        point_edge_distance_gradient(
+            p, e0, e1, grad, point_edge_distance_type(p, e0, e1));
+        break;
+
     default:
         throw std::invalid_argument(
             "Invalid distance type for point-edge distance gradient!");
     }
-}
-
-/// @brief Compute the hessian of the distance between a point and edge.
-/// @note The distance is actually squared distance.
-/// @param[in] p The point.
-/// @param[in] e0 The first vertex of the edge.
-/// @param[in] e1 The second vertex of the edge.
-/// @param[out] hess The hessian of the distance wrt p, e0, and e1.
-template <
-    typename DerivedP,
-    typename DerivedE0,
-    typename DerivedE1,
-    typename DerivedHess>
-void point_edge_distance_hessian(
-    const Eigen::MatrixBase<DerivedP>& p,
-    const Eigen::MatrixBase<DerivedE0>& e0,
-    const Eigen::MatrixBase<DerivedE1>& e1,
-    Eigen::PlainObjectBase<DerivedHess>& hess)
-{
-    assert(p.size() == 2 || p.size() == 3);
-    assert(e0.size() == 2 || e0.size() == 3);
-    assert(e1.size() == 2 || e1.size() == 3);
-
-    return point_edge_distance_hessian(
-        p, e0, e1, point_edge_distance_type(p, e0, e1), hess);
 }
 
 /// @brief Compute the hessian of the distance between a point and edge.
@@ -172,8 +116,8 @@ void point_edge_distance_hessian(
     const Eigen::MatrixBase<DerivedP>& p,
     const Eigen::MatrixBase<DerivedE0>& e0,
     const Eigen::MatrixBase<DerivedE1>& e1,
-    const PointEdgeDistanceType dtype,
-    Eigen::PlainObjectBase<DerivedHess>& hess)
+    Eigen::PlainObjectBase<DerivedHess>& hess,
+    const PointEdgeDistanceType dtype = PointEdgeDistanceType::AUTO)
 {
     int dim = p.size();
     assert(e0.size() == dim);
@@ -200,6 +144,11 @@ void point_edge_distance_hessian(
 
     case PointEdgeDistanceType::P_E:
         point_line_distance_hessian(p, e0, e1, hess);
+        break;
+
+    case PointEdgeDistanceType::AUTO:
+        point_edge_distance_hessian(
+            p, e0, e1, hess, point_edge_distance_type(p, e0, e1));
         break;
 
     default:
