+++ src/ipc/utils/save_obj.cpp
@@ -1,5 +1,9 @@
 #include "save_obj.hpp"
 
+#include <ipc/candidates/edge_vertex.hpp>
+#include <ipc/candidates/edge_edge.hpp>
+#include <ipc/candidates/face_vertex.hpp>
+#include <ipc/candidates/edge_face.hpp>
 #include <ipc/utils/eigen_ext.hpp>
 
 namespace ipc {
